灰度bug

+++ Core/Inc/logic.h
@@ -9,6 +9,7 @@ extern "C" {
 #include "oled.h"
 #include "motor.h"
 #include "gpio.h"
+#include <math.h>
 #include <stdint.h>
 
 #define PWM_Period_Motor 100
@@ -34,7 +35,7 @@ extern "C" {
      * @param direction (uint8_t) 拐弯方向(1是右拐2是左拐)
     */
     void Turn_round(uint8_t base_speed,uint8_t rad_factor,uint8_t direction);
-    long Gray_Offset_value(void);
+    int Gray_Offset_value(void);
     void Sandwich_function_01(uint8_t base_speed);
